weaviate/weaviate_service/requirements.txt
geopy
pandas
fastapi
uvicorn
pydantic
python-dotenv
weaviate-client